[OrderService] Moving directory files
