+++ OrderService/Domain/Enums/PaymentType.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain.Models.Enums
+﻿namespace Domain.Enums
 {
     public enum PaymentType
     {
