--- OrderService/Domain/Models/DTOs/BusMessage.cs
@@ -1,4 +1,0 @@
-﻿namespace Domain.Models.DTOs
-{
-    public abstract class BusMessage { }
-}
