+++ OrderService/Domain/Services/Bus/Messages/BusMessage.cs
@@ -1,0 +1,7 @@
+﻿namespace Domain.Services.Bus.Messages
+{
+    public abstract record BusMessage
+    {
+        public abstract string ToJson();
+    }
+}
